+++ docs/start/changelog.rst
@@ -9,6 +9,7 @@ ChangeLog
     * Made the falling animation much smoother. Many thanks to @mic159!
     * Changed the ``hue_ranges`` option of the tile_falling animation to
       ``line_hues`` and the ``line_tip_hue`` option to ``line_tip_hues``
+    * Added tile_balls tile animation
 
 0.9.4 - 3 Jan 2019
     * Added get_tile_positions action